+++ src/name.rs
@@ -36,7 +36,7 @@ use crate::namespace::NS_NO_PREFIX;
 /// `Name` implements some `From` instances for conversion from strings and tuples. For example:
 ///
 /// ```rust
-/// # use xml::name::Name;
+/// # use xml_no_std::name::Name;
 /// let n1: Name = "p:some-name".into();
 /// let n2: Name = ("p", "some-name").into();
 ///